+++ railties/test/generators/app_generator_test.rb
@@ -304,6 +304,10 @@ class AppGeneratorTest < Rails::Generators::TestCase
     run_generator [destination_root, "--skip-active-storage"]
 
     assert_no_gem "image_processing"
+
+    assert_file "Dockerfile" do |content|
+      assert_no_match(/libvips/, content)
+    end
   end
 
   def test_app_update_does_not_generate_active_storage_contents_when_skip_active_storage_is_given
@@ -744,6 +748,8 @@ class AppGeneratorTest < Rails::Generators::TestCase
     assert_file ".gitattributes" do |content|
       assert_no_match(/yarn\.lock/, content)
     end
+
+    assert_no_file ".node-version"
   end
 
   def test_webpack_option
@@ -763,6 +769,18 @@ class AppGeneratorTest < Rails::Generators::TestCase
 
     assert_equal 1, webpacker_called, "`javascript:install:webpack` expected to be called once, but was called #{webpacker_called} times."
     assert_gem "jsbundling-rails"
+
+    assert_file "Dockerfile" do |content|
+      assert_match(/yarn/, content)
+    end
+
+    assert_file ".node-version" do |content|
+      if ENV["NODE_VERSION"]
+        assert_match(/#{ENV["NODE_VERSION"]}/, content)
+      else
+        assert_match(/\d+\.\d+\.\d+/, content)
+      end
+    end
   end
 
   def test_esbuild_option
@@ -1014,6 +1032,15 @@ class AppGeneratorTest < Rails::Generators::TestCase
 
     assert_file ".dockerignore" do |content|
       assert_match(/config\/master\.key/, content)
+    end
+  end
+
+  def test_dockerfile
+    run_generator
+
+    assert_file "Dockerfile" do |content|
+      assert_match(/libvips/, content)
+      assert_no_match(/yarn/, content)
     end
   end